visual enhancements and SDL3 migration and event bubbling
made text and buttons smaller, which made the app better visually.
also migrated the app to SDL3, it is starting to look better already.
also added event bubbling so escape and enter affect dialogs

+++ src/SDL_Framework/SVGRasterizer.hpp
@@ -10,7 +10,7 @@ class SVGRasterizer
 {
 public:
 	SVGRasterizer(std::string path, int width, int height);
-	bool Draw(SDL_Renderer* renderer, int x, int y);
+	bool Draw(SDL_Renderer* renderer, float x, float y);
 	void SetSize(int width, int height);
 	std::optional<SDL_FPoint>  GetSVGSize() const;
 	void SetSVGPath(std::string path);
